Syntax fixes

+++ ssgan.py
@@ -248,11 +248,11 @@ print(netD)
 d_criterion = nn.BCEWithLogitsLoss()
 d_gan_criterion = nn.CrossEntropyLoss()
 fixed_noise = torch.FloatTensor(batch_size, nz, 1, 1).normal_(0, 1)
-fixed_noise = _to_var(fixed_noise).long()
+fixed_noise = _to_var(fixed_noise)
 
 
-d_gan_labels_real = torch.LongTensor(batch_size)
-d_gan_labels_fake = torch.LongTensor(batch_size)
+d_gan_labels_real = torch.FloatTensor(batch_size)
+d_gan_labels_fake = torch.FloatTensor(batch_size)
 
 real_label = 1
 fake_label = 0
@@ -275,7 +275,7 @@ for epoch in range(num_epochs):
     for i, data in enumerate(diabetes_loader_train):
         diabetes_data, diabetes_labels = data
         diabetes_data = _to_var(diabetes_data)
-        diabetes_labels = _to_var(diabetes_labels).long().squeeze()
+        diabetes_labels = _to_var(diabetes_labels).float().squeeze()
 
 
         ##########################
@@ -286,7 +286,7 @@ for epoch in range(num_epochs):
 
         netD.zero_grad()
         d_gan_labels_real = d_gan_labels_real.resize_as_(diabetes_labels.data.cpu().float()).uniform_(0, 0.3)
-        d_gan_labels_real_var = _to_var(d_gan_labels_real).long()
+        d_gan_labels_real_var = _to_var(d_gan_labels_real).float()
         output, d_class_logits_on_data, gan_logits_real, d_sample_features = netD(diabetes_data)
 
         supervised_loss = torch.mean(torch.abs(diabetes_labels - gan_logits_real))
@@ -304,14 +304,14 @@ for epoch in range(num_epochs):
         # Get the fake logits, real are obtained from above
 
 
-        noise = torch.LongTensor(batch_size, nz, 1, 1)
+        noise = torch.FloatTensor(batch_size, nz, 1, 1)
 
         noise.resize_(diabetes_labels.data.shape[0], nz, 1, 1).normal_(0, 1)
         noise_var = _to_var(noise)
         fake = netG(noise_var)
 
         d_gan_labels_fake = d_gan_labels_fake.resize_as_(diabetes_labels.data.cpu().float()).uniform_(0.9, 1.2)
-        d_gan_labels_fake_var = _to_var(d_gan_labels_fake).long()
+        d_gan_labels_fake_var = _to_var(d_gan_labels_fake).float()
         _, d_fake_logits_on_data, gan_logits_fake, _ = netD(fake.detach())
 
 
@@ -338,7 +338,7 @@ for epoch in range(num_epochs):
         netG.zero_grad()
 
 
-        noise = torch.LongTensor(batch_size, nz, 1, 1)
+        noise = torch.FloatTensor(batch_size, nz, 1, 1)
         noise.resize_(diabetes_labels.data.shape[0], nz, 1, 1).normal_(0, 1)
         noise_var = _to_var(noise)
         # Generate fake images
